Move cleanup logic into PR store
